Refactor button defaults into partial class
